FEAT: Add link to open instagram hashtag searching
